usermanager/userapp/serializer.py
from rest_framework import serializers


class UserSerializer(serializers.Serializer):
    name = serializers.CharField(max_length=300)
    street = serializers.CharField(max_length=300)
    city = serializers.CharField(max_length=100)
    state = serializers.CharField(max_length=100)
    date = serializers.DateField()


