requirements.txt
Django==2.1.*
django-prometheus
djangorestframework
pandas
openpyxl
xlrd
whitenoise